+++ CHANGELOG.md
@@ -5,9 +5,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 The format of this change log follows the advice given at [Keep a CHANGELOG](http://keepachangelog.com).
 
 ## [Unreleased]
+
+## [1.1.0] - 2015-10-19
 ### Added
 - `moodle-plugin-ci phpcbf` command. Re-formats code according to Moodle coding standards. This command is **not**
   supposed to be used on Travis CI, but rather locally to fix coding style problems.
+
 ### Changed
 - Commands no longer error when relevant files are not found.
 
@@ -25,4 +28,5 @@ The format of this change log follows the advice given at [Keep a CHANGELOG](htt
 - `moodle-plugin-ci shifter` command.  Run YUI Shifter on plugin YUI modules.
 - `moodle-plugin-ci csslint` command.  Lints the CSS files in the plugin.
 
-[Unreleased]: https://github.com/moodlerooms/moodle-plugin-ci/compare/1.0.0...master
+[Unreleased]: https://github.com/moodlerooms/moodle-plugin-ci/compare/1.1.0...master
+[1.1.0]: https://github.com/moodlerooms/moodle-plugin-ci/compare/1.0.0...1.1.0